this extends the benchmark a bit and improves performance by deleting an unused index in the sqlite coin store table.

As far as I can tell there are no queries that requires an index on whether a coin is spent or not. In fact, the mere existence of this index made the query planner mistakenly use it, causing a significant performance issue (as demonstrated by the benchmark).

My main attempt with this patch was to improve insertion performance, by removing the index and thus reducing the amount of work for insertions. This had a small effect in insertion performance, but more noticeable impact on the lookup speed of coins by name, excluding spent coins. The problem was (presumably) that sqlite chose to use the `coin_spent`  index *first*, followed by looking up the coin ID. By removing the index, the query instead looks up the coin ID first, and then discards it if the coin is spent, this is significantly faster.

I ran this benchmark on a fast machine with an SSD (Apple M1) as well as an older machine with spinning disk (Ubuntu amd64).

# SSD

Before:

```
```

After:

```
```

# Spinning disk

Before:

```
MOSTLY ADDITIONS, time: 51.8249s additions: 400400 removals: 20000
MOSTLY REMOVALS, time: 6.6823s additions: 600 removals: 140000
FULLBLOCKS, time: 76.4455s additions: 400400 removals: 400000
GET RECORDS BY NAMES with spent, time: 1.2238s 200 lookups found 40000 coins in total
GET RECORDS BY NAMES without spent, time: 70.1104s 200 lookups found 20730 coins in total
GET COINS REMOVED AT HIGHT, time: 22.4741s 800 blocks, found 580000 coins in total
```

After:

```
MOSTLY ADDITIONS, time: 51.6045s additions: 400400 removals: 20000
MOSTLY REMOVALS, time: 4.5481s additions: 600 removals: 140000
FULLBLOCKS, time: 71.1832s additions: 400400 removals: 400000
GET RECORDS BY NAMES with spent, time: 1.2226s 200 lookups found 40000 coins in total
GET RECORDS BY NAMES without spent, time: 0.9238s 200 lookups found 20732 coins in total
GET COINS REMOVED AT HEIGHT, time: 22.4992s 800 blocks, found 580000 coins in total
```