@mariano54 2021-09-30
I originally thought the index was necessary for querying `spent=0`, but it's not. This index has just caused performance issues, and it's good to remove it.