@arvidn 2021-10-01
One thing to keep in mind is that if someone downgraded chia from this patch, they'll get an unpleasant surprise on startup.

It might be more conservative to first remove the statement that creates the index, and then a few versions later actually remove it.