@emlowe 2021-10-05
I think Bill and I were more comfortable with removing the create index command for this and then some subsequent release dropping the index. As mentioned this helps to avoid the aforementioned unpleasantness.